MS data backend for mgf files

The MsBackendMgf class supports import of MS/MS spectra data from files in Mascot Generic Format (mgf) files. After initial import, the full MS data is kept in memory. MsBackendMgf extends the MsBackendDataFrame() backend directly and supports thus the applyProcessing() function to make data manipulations persistent. The backend does however not support export to mgf files yet.

New objects are created with the MsBackendMgf function. The backendInitialize method has to be subsequently called to initialize the object and import MS/MS data from (one or more) mgf files. Optional parameter nonStop allows to specify whether the import returns with an error if one of the xml files lacks required data, such as mz and intensity values (default nonStop = FALSE), or whether only affected file(s) is(are) skipped and a warning is shown (nonStop = TRUE). Note that any other error (such as xml import error) will abort import regardless of parameter nonStop.

# S4 method for MsBackendMgf
backendInitialize(object, files, nonStop = FALSE, ..., BPPARAM = bpparam())

MsBackendMgf()

Arguments

object
Instance of MsBackendMgf class.

files
character with the (full) file name(s) of the mgf file(s) from which MS/MS data should be imported.

nonStop
logical(1) whether import should be stopped if an xml file does not contain all required fields. Defaults to nonStop = FALSE.

...
Currently ignored.

BPPARAM
Parameter object defining the parallel processing setup to import data in parallel. Defaults to BPPARAM = bpparam(). See bpparam() for more information.

Examples

## Create an MsBackendHmdbXml backend and import data from test xml files.
fls <- dir(system.file("extdata", package = "MsBackendMgf"),
    full.names = TRUE, pattern = "mgf$")
be <- backendInitialize(MsBackendMgf(), fls)

#> Start data import from 2 files ... 

#> done

be

#> MsBackendMgf with 7 spectra
#>     msLevel     rtime scanIndex
#>   <integer> <numeric> <integer>
#> 1         2   1028.00      2162
#> 2         2   1117.00      2406
#> 3         2   1127.00      2432
#> 4         2   2678.94         0
#> 5         2   2373.51         0
#> 6         2   2511.03         0
#> 7         2      0.00         0
#>  ... 17 more variables/columns.

be$msLevel

#> [1] 2 2 2 2 2 2 2

be$intensity

#> NumericList of length 7
#> [[1]] 753.738 385.376 315.441 413.206 ... 3038.73 2016.43 1146.04 704.175
#> [[2]] 1228.93 1424.66 1550.9 1455.45 ... 7380.41 4960.92 5743.83 1780.76
#> [[3]] 1340.44 1714.76 1938.82 1450.36 2019 ... 5323.02 2265.43 4768.14 1532.12
#> [[4]] 81011.57 4123.349 4006.9321 66933.17 ... 22042.248 18096.48 12666.438
#> [[5]] 14.1766004562378 18.5806427001953 ... 22.7096385955811 14.864013671875
#> [[6]] 1748.57495117188 8689.9951171875 ... 2907.08422851562 2663.30908203125
#> [[7]] 65.219513 178.758606 13.01786 119.898499 ... 22.05921 30.57095 14.11111

be$mz

#> NumericList of length 7
#> [[1]] 102.0548 103.00494 103.03531 ... 1388.58691 1405.59729 1406.57666
#> [[2]] 101.07074 102.05486 103.00227 ... 1331.56726 1348.58496 1349.59241
#> [[3]] 102.05556 103.00014 115.05058 ... 1333.599 1334.61304 1335.64368
#> [[4]] 101.07122 109.68925 115.86999 120.0811 ... 1260.6073 1261.614 1272.6572
#> [[5]] 130.164459228516 144.150299072266 ... 1019.23852539062 1020.52404785156
#> [[6]] 110.070594787598 120.080627441406 ... 887.756652832031 998.447387695312
#> [[7]] 51.022404 57.033543 57.060638 ... 636.130188 660.481445 753.358521